fix(sheet): move mutation code to column manager
